Fit board on screen and matched scaling of grid and piece. Causes problem with hit function and movement, will fix shortly

+++ app/src/main/java/edu/msu/weath151/connect4/Piece.java
@@ -67,7 +67,7 @@ public class Piece {
         return piece;
     }
 
-    public void onDraw(Canvas canvas, int marginX, int marginY,
+    public void onDraw(Canvas canvas, float marginX, float marginY,
                        int puzzleSize, float scaleFactor){
 
         canvas.save();
